Fixed compiler warnings.
git-svn-id: file:///home/xapian-svn/xapian/trunk@3190 6b2f1b92-a4f3-0310-943f-de3cbe41748b

+++ xapian-core/backends/quartz/quartz_metafile.cc
@@ -51,7 +51,7 @@ static std::string encode_version(unsigned int version)
 {
     std::string data;
 
-    for (int i=0; i<sizeof(metafile_version); ++i) {
+    for (size_t i=0; i<sizeof(metafile_version); ++i) {
 	data += (char)(version & 0xff);
 	version >>= 8;
     }
